requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.0.4
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
pytesseract==0.3.10
joblib==1.3.2
cryptography==41.0.4
psycopg2-binary==2.9.7
